# event-booking-swagger.yml
swagger: '2.0'
info:
  version: 1.0.0
  title: Event Booking API
  description: This API allows customers to create, find and reserve tickets for events, view and manage their reservations, and be notified before the event kickoff.
host: localhost:8080
schemes:
  - http
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
paths:
  /users:
    post:
      summary: Create a new user.
      description: This endpoint allows customers to create a new user.
      parameters:
        - in: body
          name: user
          description: The user to create.
          schema:
            $ref: '#/definitions/User'
      responses:
        '201':
          description: Created
        '400':
          description: Bad request
      tags:
        - users
  /auth:
    post:
      summary: Authenticate a user.
      description: This endpoint allows users to authenticate and receive a Bearer token.
      parameters:
        - in: body
          name: credentials
          description: The user's credentials.
          schema:
            $ref: '#/definitions/Credentials'
      responses:
        '200':
          description: OK
          headers:
            Authorization:
              type: string
        '401':
          description: Unauthorized
      tags:
        - users
  /events:
    post:
      summary: Create a new event.
      description: This endpoint allows customers to create a new event.
      parameters:
        - in: body
          name: event
          description: The event to create.
          schema:
            $ref: '#/definitions/EventRequestDTO'
      responses:
        '201':
          description: Created
          schema:
            type: object
            properties:
              eventId:
                type: integer
                format: string
        '400':
          description: Bad request
        '401':
          description: Unauthorized
      tags:
        - events
      security:
        - Bearer: []
    get:
      summary: Get all events or search for events.
      description: This endpoint allows customers to retrieve all events or search for events by name, date range or category.
      parameters:
        - in: query
          name: name
          description: The name of the event to search for.
          type: string
        - in: query
          name: startDate
          description: The start date of the date range to search for events in (inclusive).
          type: string
          format: date
        - in: query
          name: endDate
          description: The end date of the date range to search for events in (inclusive).
          type: string
          format: date
        - in: query
          name: category
          description: The category of the event to search for.
          type: string
          enum:
            - Concert
            - Conference
            - Game
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/EventResponseDTO'
        '404':
          description: Not found
      tags:
        - events
  /events/{eventId}/tickets:
    post:
      summary: Reserve tickets for an event.
      description: This endpoint allows customers to reserve tickets for an event.
      parameters:
        - in: path
          name: eventId
          description: The ID of the event to reserve tickets for.
          required: true
          type: integer
          format: string
        - in: body
          name: ticketRequest
          description: The ticket reservation request.
          schema:
            $ref: '#/definitions/TicketRequest'
      responses:
        '201':
          description: Created
        '400':
          description: Bad request
        '401':
          description: Unauthorized
      tags:
        - tickets
      security:
        - Bearer: []
  /events/{eventId}/cancel:
    post:
      summary: Cancel a reservation for an event
      tags:
        - tickets
      description: This endpoint allows customers to cancel reservations
      parameters:
        - in: path
          name: eventId
          description: The ID of the event to reserve tickets for.
          required: true
          type: integer
          format: string
      responses:
        '201':
          description: Created
        '400':
          description: Bad request
        '401':
          description: Unauthorized
      security:
        - Bearer: []
  /events/{eventId}/booked:
    get:
      summary: Get booked events
      tags:
        - tickets
      description: This endpoint allows customers to get their booked events
      parameters:
        - in: path
          name: eventId
          description: The ID of the event to reserve tickets for.
          required: true
          type: integer
          format: string
      responses:
        '200':
          description: OK
        '400':
          description: Bad request
        '401':
          description: Unauthorized
      security:
        - Bearer: []
definitions:
  User:
    type: object
    required:
      - name
      - email
      - password
    properties:
      name:
        type: string
        maxLength: 100
      email:
        type: string
        format: email
      password:
        type: string
        minLength: 8
  Credentials:
    type: object
    required:
      - email
      - password
    properties:
      email:
        type: string
        format: email
      password:
        type: string
        minLength: 8
  EventRequestDTO:
    type: object
    required:
      - name
      - date
      - availableAttendeesCount
      - description
      - category
    properties:
      name:
        type: string
        maxLength: 100
      date:
        type: string
        format: date
      availableAttendeesCount:
        type: integer
        minimum: 1
        maximum: 1000
      description:
        type: string
        maxLength: 500
      category:
        $ref: '#/definitions/Category'
  EventResponseDTO:
    type: object
    properties:
      id:
        type: integer
        format: string
      name:
        type: string
      date:
        type: string
        format: date
      availableAttendeesCount:
        type: integer
      description:
        type: string
      category:
        $ref: '#/definitions/Category'
  TicketRequest:
    type: object
    required:
      - attendeesCount
    properties:
      attendeesCount:
        type: integer
        minimum: 1
  Category:
    type: string
    enum:
      - Concert
      - Conference
      - Game